--- a/src/hypertrace/agent/filter/__init__.py
+++ b/src/hypertrace/agent/filter/__init__.py
@@ -9,7 +9,7 @@ class Filter(ABC):
     continues it.
     """
     @abstractmethod
-    def evaluate_url_and_headers(self, span: Span, url: str, headers: tuple) -> bool:
+    def evaluate_url_and_headers(self, span: Span, url: str, headers: dict) -> bool:
         """evaluate_url_and_headers can be used to evaluate both URL and Header"""
 
     @abstractmethod
@@ -20,7 +20,7 @@ class Filter(ABC):
 class NoopFilter(Filter):
     """NoopFilter is a filter that never blocks"""
 
-    def evaluate_url_and_headers(self, span: Span, url: str, headers: tuple) -> bool:
+    def evaluate_url_and_headers(self, span: Span, url: str, headers: dict) -> bool:
         return False
 
     def evaluate_body(self, span: Span, body) -> bool: