## Description
The goal here is to handle some low hanging fruit to improve performance: 

1.) Less `logger.debug("some message %s", str(a_span))`
Even though the log messages dont have the stringified objects interpolated when not in debug logging the `str` cast call still occurs.
(This does get costly when doing it per req for every header + body)
 
2.) headers tuple → dict
We were adding some overhead in multiple instrumentations because we were converting a header dict to a list of tuples. Using a dict instead gives us a few benefits: 

- content type lookup can be simple key lookup, no need to iterate over all headers
- Saves us a conversions from dict to tuple


left = current released version, right = this PR
<img width="1131" alt="Screen Shot 2021-09-24 at 5 31 24 PM" src="https://user-images.githubusercontent.com/15878949/134742710-1710485b-043e-4c2c-b07a-a578fe9079b7.png">

The profiler data below has been filtered to just show hypertrace calls. 
The important detail is that this PR has lower top level ttot values than the current release:

ex: 
hypertrace_before_request 0.13 vs 0.19 

(This PR + 100 req)

```yaml
name                                  ncall  tsub      ttot      tavg      
..__.py:48 hypertrace_before_request  100    0.007830  0.139217  0.001392
..InstrumentedFlask._generic_handler  199    0.005692  0.101931  0.000512
..entedFlask.generic_request_handler  100    0.001194  0.071479  0.000715
..trumentedFlask.add_headers_to_span  199    0.002609  0.048455  0.000243
..t__.py:90 hypertrace_after_request  99     0.002903  0.047636  0.000481
..nstrumentedFlask.lowercase_headers  199    0.001573  0.035597  0.000179
..entation/__init__.py:76 <dictcomp>  199    0.004087  0.034024  0.000171
..ntedFlask.generic_response_handler  99     0.000918  0.033897  0.000342
..gent/filter/registry.py:13 __new__  100    0.004116  0.004442  0.000044
..ask.eligible_based_on_content_type  199    0.000993  0.001368  0.000007
..strumentedFlask.grab_first_n_bytes  99     0.000678  0.001348  0.000014
..eInstrumentedFlask.check_body_size  99     0.000495  0.000670  0.000007
..r/registry.py:18 Registry.__init__  100    0.000427  0.000649  0.000006
..istry.py:28 Registry.apply_filters  100    0.000317  0.000317  0.000003
```

Current released version + 100 Requests

```yaml
..__.py:48 hypertrace_before_request  100    0.009801  0.198750  0.001988
..entedFlask.generic_request_handler  100    0.014139  0.116980  0.001170
..t__.py:92 hypertrace_after_request  99     0.003852  0.074692  0.000754
..ntedFlask.generic_response_handler  99     0.009162  0.050550  0.000511
..ntation/__init__.py:146 <listcomp>  100    0.002642  0.027981  0.000280
..strumentedFlask.grab_first_n_bytes  99     0.000691  0.001951  0.000020
..eInstrumentedFlask.check_body_size  99     0.000835  0.001260  0.000013
..ntation/__init__.py:211 <listcomp>  99     0.000691  0.001061  0.000011
..dFlask.is_interesting_content_type  99     0.000480  0.001057  0.000011
..gent/filter/registry.py:13 __new__  100    0.000505  0.000754  0.000008
..r/registry.py:18 Registry.__init__  100    0.000438  0.000659  0.000007
..istry.py:28 Registry.apply_filters  100    0.000370  0.000370  0.000004
..dFlask.get_process_request_headers  100    0.000280  0.000280  0.000003
..Flask.get_process_response_headers  99     0.000270  0.000270  0.000003
..ntedFlask.get_process_request_body  100    0.000263  0.000263  0.000003
..tedFlask.get_process_response_body  99     0.000254  0.000254  0.000003
..nstrumentedFlask.get_max_body_size  99     0.000236  0.000236  0.000002
```